app/requirements.txt
django==4.0.1
djangorestframework==3.13.1
python-dotenv==0.19.2
psycopg2-binary==2.9.3
pytest-django==4.5.2
pytest==6.2.5